Issue #3042682 by driverok, jkswoods, mikeegoulding, Matroskeen, Sudeepthi Peteti, bircher: Drupal 9 Deprecated Code Report

+++ src/Plugin/ConfigFilter/SplitFilter.php
@@ -12,6 +12,7 @@ use Drupal\Core\Database\Connection;
 use Drupal\Core\DependencyInjection\DependencySerializationTrait;
 use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Drupal\Component\PhpStorage\FileStorage as PhpFileStorage;
+use Drupal\Component\FileSecurity\FileSecurity;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
@@ -410,7 +411,12 @@ class SplitFilter extends ConfigFilterBase implements ContainerFactoryPluginInte
       if (file_exists($directory) && is_writable($directory)) {
         $htaccess_path = rtrim($directory, '/\\') . '/.htaccess';
         if (!file_exists($htaccess_path)) {
-          file_put_contents($htaccess_path, PhpFileStorage::htaccessLines(TRUE));
+          if (class_exists(FileSecurity::class)) {
+            file_put_contents($htaccess_path, FileSecurity::htaccessLines(TRUE));
+          }
+          else {
+            file_put_contents($htaccess_path, PhpFileStorage::htaccessLines(TRUE));
+          }
           @chmod($htaccess_path, 0444);
         }
       }
